# [xamlc] remove `LoadInSeparateAppDomain` and `static` state