--- a/src/Controls/src/Build.Tasks/CompiledConverters/ColumnDefinitionCollectionTypeConverter.cs
+++ b/src/Controls/src/Build.Tasks/CompiledConverters/ColumnDefinitionCollectionTypeConverter.cs
@@ -20,19 +20,19 @@ namespace Microsoft.Maui.Controls.XamlC
 				var parts = value.Split(',');
 
 				yield return Create(Ldc_I4, parts.Length);
-				yield return Create(Newarr, module.ImportReference(("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "ColumnDefinition")));
+				yield return Create(Newarr, module.ImportReference(context.Cache, ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "ColumnDefinition")));
 				for (var i = 0; i < parts.Length; i++)
 				{
 					yield return Create(Dup);
 					yield return Create(Ldc_I4, i);
 					foreach (var instruction in gridlengthconverter.ConvertFromString(parts[i], context, node))
 						yield return instruction;
-					yield return Create(Newobj, module.ImportCtorReference(
+					yield return Create(Newobj, module.ImportCtorReference(context.Cache,
 							type: ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "ColumnDefinition"),
 							parameterTypes: new[] { ("Microsoft.Maui", "Microsoft.Maui", "GridLength") }));
 					yield return Create(Stelem_Ref);
 				}
-				yield return Create(Newobj, module.ImportCtorReference(
+				yield return Create(Newobj, module.ImportCtorReference(context.Cache,
 						type: ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "ColumnDefinitionCollection"),
 						paramCount: 1));
 				yield break;