--- a/src/Controls/src/Build.Tasks/CompiledConverters/EasingTypeConverter.cs
+++ b/src/Controls/src/Build.Tasks/CompiledConverters/EasingTypeConverter.cs
@@ -25,7 +25,7 @@ namespace Microsoft.Maui.Controls.XamlC
 			var assemblyTypeInfo = ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", nameof(Easing));
 
 			var module = context.Body.Method.Module;
-			var fieldReference = module.ImportFieldReference(assemblyTypeInfo, value, isStatic: true, caseSensitive: false);
+			var fieldReference = module.ImportFieldReference(context.Cache, assemblyTypeInfo, value, isStatic: true, caseSensitive: false);
 
 			if (fieldReference != null)
 			{