--- a/src/Controls/src/Build.Tasks/CompiledConverters/RowDefinitionCollectionTypeConverter.cs
+++ b/src/Controls/src/Build.Tasks/CompiledConverters/RowDefinitionCollectionTypeConverter.cs
@@ -20,7 +20,7 @@ namespace Microsoft.Maui.Controls.XamlC
 				var parts = value.Split(',');
 
 				yield return Create(Ldc_I4, parts.Length);
-				yield return Create(Newarr, module.ImportReference(("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "RowDefinition")));
+				yield return Create(Newarr, module.ImportReference(context.Cache, ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "RowDefinition")));
 				for (var i = 0; i < parts.Length; i++)
 				{
 					yield return Create(Dup);
@@ -28,11 +28,13 @@ namespace Microsoft.Maui.Controls.XamlC
 					foreach (var instruction in gridlengthconverter.ConvertFromString(parts[i], context, node))
 						yield return instruction;
 					yield return Create(Newobj, module.ImportCtorReference(
+							context.Cache,
 							type: ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "RowDefinition"),
 							parameterTypes: new[] { ("Microsoft.Maui", "Microsoft.Maui", "GridLength") }));
 					yield return Create(Stelem_Ref);
 				}
 				yield return Create(Newobj, module.ImportCtorReference(
+						context.Cache,
 						type: ("Microsoft.Maui.Controls", "Microsoft.Maui.Controls", "RowDefinitionCollection"),
 						paramCount: 1));
 				yield break;