… src/Controls/src/Build.Tasks/CompiledConverters/RDSourceTypeConverter.cs
… src/Controls/src/Build.Tasks/CompiledConverters/StrokeShapeTypeConverter.cs
… src/Controls/src/Build.Tasks/CompiledConverters/ThicknessTypeConverter.cs
… src/Controls/src/Build.Tasks/CompiledConverters/TypeTypeConverter.cs
… src/Controls/src/Build.Tasks/CompiledConverters/UriTypeConverter.cs
… src/Controls/src/Build.Tasks/CompiledMarkupExtensions/ArrayExtension.cs
… src/Controls/src/Build.Tasks/CompiledMarkupExtensions/DataTemplateExtension.cs
… src/Controls/src/Build.Tasks/CompiledMarkupExtensions/StaticExtension.cs
… src/Controls/src/Build.Tasks/CompiledMarkupExtensions/TypeExtension.cs
… src/Controls/src/Build.Tasks/CompiledValueProviders/SetterValueProvider.cs
… src/Controls/src/Build.Tasks/CompiledValueProviders/StyleSheetProvider.cs
… src/Controls/src/Build.Tasks/CreateObjectVisitor.cs
… src/Controls/src/Build.Tasks/DebugXamlCTask.cs
… src/Controls/src/Build.Tasks/ExpandMarkupsVisitor.cs
… src/Controls/src/Build.Tasks/ILContext.cs
… src/Controls/src/Build.Tasks/MethodDefinitionExtensions.cs
… src/Controls/src/Build.Tasks/ModuleDefinitionExtensions.cs
… src/Controls/src/Build.Tasks/NodeILExtensions.cs
… src/Controls/src/Build.Tasks/SetFieldVisitor.cs
… src/Controls/src/Build.Tasks/SetNamescopesAndRegisterNamesVisitor.cs
… src/Controls/src/Build.Tasks/SetPropertiesVisitor.cs
… src/Controls/src/Build.Tasks/SetResourcesVisitor.cs
… src/Controls/src/Build.Tasks/TypeDefinitionExtensions.cs
… src/Controls/src/Build.Tasks/TypeReferenceExtensions.cs
… src/Controls/src/Build.Tasks/VariableDefinitionExtensions.cs
… src/Controls/src/Build.Tasks/XamlCTask.cs
… src/Controls/src/Build.Tasks/XamlCache.cs
… src/Controls/src/Build.Tasks/XamlTask.cs
… src/Controls/src/Build.Tasks/XmlTypeExtensions.cs
… src/Controls/tests/Xaml.UnitTests/CecilExtensionsTests.cs