@jonathanpeppers 2022-12-20
I'm going to merge -- it should be for .NET 8 only, so I should have a good amount of time to address any problems. Thanks.